organize classes 01
